+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.15
 * Add timeout option
+* Add AEM Password Reset Activator support
 
 ### 1.0.14
 * Switch from deleteAgent to postAgent operation for all agents delete action